Refactor for Clarity
